requirements.txt
aiohttp
APScheduler
async-timeout
chardet
discord.py
emoji
multidict
pytz
six
tzlocal
websockets
black
pyyaml
google-cloud-firestore
tabulate
python-dotenv

# Dev
pre-commit